backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
django-filter>=24.3,<25
psycopg[binary]>=3.2,<4
celery>=5.4,<6
redis>=5.2,<6
python-dotenv>=1.0,<2
inflection>=0.5,<1
uritemplate>=4.2,<5
